% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prevalence_calc.R
\name{a_priori_est_scc}
\alias{a_priori_est_scc}
\title{A priori estimations for SCC}
\usage{
a_priori_est_scc(
  baserate_hp = 211/1194,
  devices = "HP",
  switch_to_target = 3/4,
  min_number,
  min_prob = 0.8,
  tolerance = as.integer(0)
)
}
\arguments{
\item{baserate_hp}{Sets the (estimated) prevalence of headphones in the
target population as a number between 0 and 1. Defaults to the unbiased
prevalence B of 0.1767 from \insertCite{HALT_2;textual}{HALT}.}

\item{devices}{Sets the desired playback device. Possible settings are
\code{"HP"} for headphones or \code{"LS"} for loudspeakers.}

\item{switch_to_target}{Sets the (estimated) switching prevalence.
The switching prevalence describes the probability that a participant who
indicates the use of a device other than the target device actually switches
to the target device after being prompted to do so.}

\item{min_number}{minimum number of participants passing the test procedure
and using the target devices}

\item{min_prob}{minimum probability (equal to or greater than 0.6, less than
1) for the event that at least \code{min_number} participants pass the test
procedure and use the target devices.}

\item{tolerance}{(non-negative integer) defaults to \code{0}. A value of 0
searches for the screening procedures with the minimum sample size.
If set to a value > 0 the function returns the screening procedures whose
sample sizes exceed the minimum sample size by at most this value.}
}
\description{
A priori estimations for SCC
}
\section{Model}{

The function uses the Normal approximation of the Binomial distribution with
continuity correction. For details, see \insertCite{HALT_2;textual}{HALT}.
}

\references{
\insertAllCited{}
}
