% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a_priori_estimation.R
\name{a_priori_estimation}
\alias{a_priori_estimation}
\title{A priori estimaion}
\usage{
a_priori_estimation(
  screening_strat,
  devices = "HP",
  baserate_hp = 211/1194,
  switch_to_target = NA,
  min_number,
  min_prob = 0.8,
  tolerance = as.integer(0)
)
}
\arguments{
\item{screening_strat}{three-letter lower-case abbreviation of the screening
strategy. See \code{\link{screening_strategies}}}

\item{devices}{Sets the desired playback device. Possible settings are
\code{"HP"} for headphones or \code{"LS"} for loudspeakers.}

\item{baserate_hp}{Sets the (estimated) prevalence of headphones in the
target population as a number between 0 and 1. Defaults to the unbiased
prevalence B of 0.1767 from \insertCite{HALT_2;textual}{HALT}.}

\item{switch_to_target}{Sets the (estimated) switching prevalence.
The switching prevalence describes the probability that a participant who
indicates the use of a device other than the target device actually switches
to the target device after being prompted to do so.}

\item{min_number}{minimum number of participants \code{k} who are in the
final sample and should have used the correct device.
When you use \code{min_number} you cannot use \code{min_prob}, i.e.
\code{min_prob = NULL}.}

\item{min_prob}{(greater than 0, less than 1) for the event that
at least an unknown number of participants \code{k} who are in the final
sample used the correct device.
When you use \code{min_prob} you cannot use \code{min_number}, i.e.
\code{min_number = NULL}.}
}
\value{
A tibble (data frame) with the characteristics of the test
procedure(s) and the attribute \code{explanation}.
This attribute is intended as an explanatory text containing a
probabilistic statement for the test procedure requiring the smallest sample
size.
}
\description{
This function provides estimations for the sample size required for the
event that at least a minimum number of participants used the desired
devices to have a specified minimum probability. By default, the function
returns the screening procedure that requires the smallest sample size for
this event. To return screening procedures with similar required sample
sizes use the parameter \code{tolerance}.
}
\section{Model}{

The function uses the Normal approximation of the Binomial distribution with
continuity correction. For details, see \insertCite{HALT_2;textual}{HALT}.
}

\references{
\insertAllCited{}
}
