% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a_priori_estimation.R
\name{a_priori_explanation}
\alias{a_priori_explanation}
\title{A priori explanation}
\usage{
a_priori_explanation(
  screening_strat,
  devices,
  baserate_hp = 211/1194,
  switch_to_target = NA,
  min_number,
  min_prob = 0.8,
  test_method
)
}
\description{
This function provides an explanatory text for an a priori estimation.
It is used inside \code{\link{a_priori_estimation}}.
}
