% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_config.R
\name{auto_config}
\alias{auto_config}
\title{Automatic configuration for the HALT module}
\usage{
auto_config(
  baserate_hp = 211/1194,
  devices = "HP",
  use_scc = FALSE,
  loop_exclude = 5L,
  lr_img_exclude = TRUE,
  lr_audio_exclude = TRUE,
  devices_exclude = TRUE
)
}
\arguments{
\item{baserate_hp}{Sets the (estimated) prevalence of headphones in the
target population as a number between 0 and 1. Defaults to the unbiased
prevalence B of 0.1767.
from Wycisk et al.(2020).}

\item{devices}{Sets the desired playback device. Possible settings are
\code{"HP"} for headphones or \code{"LS"} for loudspeakers.}

\item{use_scc}{(boolean, default = F) Flag whether "Splice, Convice. Compare" page shall be shown.}

\item{loop_exclude}{(integer, default = 5) Number of loops for item po2.}

\item{lr_img_exclude}{(boolean, default = T) Flag, if wrong answer on left-right image question shall lead to exclusiogn.}

\item{lr_audio_exclude}{(boolean, default = T) Flag, if wrong answer on left-right audio question shall lead to exclusiogn.}

\item{devices_exclude}{(boolean, default = T) Flag, if wrong answer on left-right device question shall lead to exclusiogn.}
}
\description{
This functions determines the parameters of the psychTestR HALT function
based on predefined criteria for the specified application.
}
\references{
\insertRef{HALTpaper}{HALT}
}
