% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation_keys.R
\name{evaluation_keys}
\alias{evaluation_keys}
\alias{combination_methods}
\title{Evaluation Keys / Test Combinations}
\usage{
evaluation_keys()

combination_methods()
}
\description{
This functions returns all screening test (combination) methods available in
HALT with their respective evaluation key (or method code).
For details, see \insertCite{HALT_2;textual}{HALT}.
}
\references{
\insertAllCited{}
}
