% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_config.R
\name{export_config}
\alias{export_config}
\title{Export config as csv}
\usage{
export_config(config, file = "")
}
\arguments{
\item{config}{object of class HALT_config}

\item{file}{character string naming a file open for writing.}
}
\description{
This function exports a HALT config object as csv file.
}
