% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{extract_HALT_items}
\alias{extract_HALT_items}
\title{This function reads a directory of results files and extracts a tibble of HALT data}
\usage{
extract_HALT_items(result_dir)
}
\arguments{
\item{result_dir}{(character scalar) Result dir to parse}
}
\description{
This function reads a directory of results files and extracts a tibble of HALT data
}
