% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HALT.R
\name{HALT}
\alias{HALT}
\title{HALT}
\usage{
HALT(
  label = "HALT",
  config = HALT::auto_config(),
  audio_dir = "https://media.gold-msi.org/test_materials/HLT",
  dict = HALT::HALT_dict
)
}
\arguments{
\item{label}{(Character scalar) Label to give the HALT results in the output file.}

\item{config}{(object of class HALT_config) Use function auto_config() or make_config() to generate this.
See also the documentation there for further explanations.}

\item{audio_dir}{(url). The URL of the directory containing the stimuli.}

\item{dict}{The psychTestR dictionary used for internationalisation.}
}
\description{
This function defines a HALT  module for incorporation into a
psychTestR timeline.
Use this function if you want to include the HALT in a
battery of other tests, or if you want to add custom psychTestR
pages to your test timeline.
For a standalone implementation of the HALT,
consider using \code{\link{HALT_standalone}()}.
}
