% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo.R
\name{HALT_demo}
\alias{HALT_demo}
\title{Demo HALT}
\usage{
HALT_demo(
  num_items = 3L,
  feedback = NULL,
  config = HALT::auto_config(use_scc = T),
  admin_password = "demo",
  researcher_email = "",
  dict = HALT::HALT_dict,
  language = "en",
  ...
)
}
\arguments{
\item{num_items}{(Integer scalar) Number of items in the RAT test.}

\item{feedback}{(Function) Defines the feedback to give the participant
at the end of the test. Defaults to a graph-based feedback page.}

\item{admin_password}{(Scalar character) Password for accessing the admin panel.
Defaults to \code{"demo"}.}

\item{researcher_email}{(Scalar character)
If not \code{NULL}, this researcher's email address is displayed
at the bottom of the screen so that online participants can ask for help.
Defaults to \email{longgold@gold.uc.ak},
the email address of this package's developer.}

\item{dict}{The psychTestR dictionary used for internationalisation.}

\item{language}{The language you want to run your demo in.
Possible languages include English (\code{"EN"}) and German (\code{"DE"}).
The first language is selected by default}

\item{...}{Further arguments to be passed to \code{\link{RAT}()}.}
}
\description{
This function launches a demo for the HALT, which is wrapped around a RAT test.
}
