% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone.R
\name{HALT_standalone}
\alias{HALT_standalone}
\title{Standalone HALT}
\usage{
HALT_standalone(
  title = NULL,
  with_id = FALSE,
  config = HALT::auto_config(),
  admin_password = "HALTadmin",
  researcher_email = "yves.wycisk@hmtm-hannover.de",
  languages = c("en", "de"),
  dict = HALT::HALT_dict,
  audio_dir = "https://media.gold-msi.org/test_materials/HLT",
  validate_id = "auto",
  ...
)
}
\arguments{
\item{title}{(Scalar character) Title to display during testing. If \code{NULL}, \code{TESTNAME} from the dictionaly is used.}

\item{with_id}{(Logical character) Should an id page be included at the beginning.}

\item{config}{(object of class HALT_config) Use function auto_config() or make_config() to generate this.
See also the documentation there for further explanations.}

\item{admin_password}{(Scalar character) Password for accessing the admin panel.}

\item{researcher_email}{(Scalar character)
If not \code{NULL}, this researcher's email address is displayed
at the bottom of the screen so that online participants can ask for help.}

\item{languages}{(Character vector)
Determines the languages available to participants.
Possible languages include English (\code{"en"}) and German (\code{"de"}).
The first language is selected by default}

\item{dict}{The psychTestR dictionary used for internationalisation.}

\item{audio_dir}{(url). The URL of the directory containing the stimuli.}

\item{validate_id}{(Character scalar or closure) Function for validating IDs or string "auto" for default validation
which means ID should consist only of  alphanumeric characters.}

\item{...}{Further arguments to be passed to \code{\link{HALT}()}.}
}
\description{
This function launches a standalone testing session for the HALT
This can be used for data collection, either in the laboratory or online.
}
