% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_config.R
\name{import_config}
\alias{import_config}
\title{Import a csv config}
\usage{
import_config(file)
}
\arguments{
\item{file}{character string naming the csv file}
}
\description{
This function imports a csv config file as HALT config object.
}
