% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prevalence_calc.R
\name{max_hp_pv}
\alias{max_hp_pv}
\title{Maximum Headphones Predictive Value}
\usage{
max_hp_pv(baserate_hp = 211/1194)
}
\arguments{
\item{baserate_hp}{Sets the (estimated) prevalence of headphones in the
target population as a number between 0 and 1. Defaults to the unbiased
prevalence B of 0.1767 from \insertCite{HALTpaper;textual}{HALT}.}
}
\description{
This function returns a data.frame containing all tests with maximum
headphones predictive value for the specified prevalence of headphones.
}
\references{
\insertRef{HALTpaper}{HALT}
}
