% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prevalence_calc.R
\name{max_utility}
\alias{max_utility}
\title{Maximum Utility maximizing percent correct}
\usage{
max_utility(baserate_hp = 211/1194)
}
\arguments{
\item{baserate_hp}{Sets the (estimated) prevalence of headphones in the
target population as a number between 0 and 1. Defaults to the unbiased
prevalence B of 0.1767 from \insertCite{HALTPaper;textual}{HALT}.}
}
\description{
This function returns a data.frame containing all tests with maximum value
for utility maximizing percent correct for the specified prevalence of
headphones.
}
\references{
\insertRef{HALTpaper}{HALT}
}
