% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prevalence_calc.R
\name{post_hoc_est}
\alias{post_hoc_est}
\title{Post hoc estimations}
\usage{
post_hoc_est(
  combination_method,
  A_threshold,
  B_threshold,
  C_threshold,
  baserate_hp,
  devices,
  samplesize,
  min_number = NULL,
  min_prob = NULL
)
}
\arguments{
\item{combination_method}{Number (1 to 18) corresponding to the test method.}

\item{A_threshold}{(scalar integer) Threshold for Test A (1 to 6).}

\item{B_threshold}{(scalar integer) Threshold for Test B (1 to 6).}

\item{C_threshold}{(scalar integer) Threshold for Test C (1 to 6).}

\item{baserate_hp}{Sets the (estimated) prevalence of headphones in the
target population as a number between 0 and 1. Defaults to the unbiased
prevalence B of 0.1767 from \insertCite{HALT_2;textual}{HALT}.}

\item{devices}{Sets the desired playback device. Possible settings are
\code{"HP"} for headphones or \code{"LS"} for loudspeakers.}

\item{samplesize}{number of participants classified as users of the target
device}

\item{min_number}{minimum number of participants \code{k} who passed the
test procedure and used the correct device}

\item{min_prob}{Probability (greater than 0, less than 1) for the event that
at least an unknown number of participants \code{k} who passed the test
procedure used the correct device}
}
\description{
This function provides probabilistic statements about the composition of a
sample after application of a certain test procedure. For this purpose a
Binomial distribution is used.
}
\details{
Given a test procedure, prevalence, and sample size the event that at least
\code{k} participants who passed the test procedure used the correct device
is considered.
The function either calculates the minimum probability for this event for a
given \code{k} or \code{k} for a given probability for this event.
}
\note{
Only one of the arguments \code{min_number} and \code{min_prob} can be
used.
}
\references{
\insertAllCited{}
}
