% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prevalence_calc.R
\name{post_hoc_est_fwr}
\alias{post_hoc_est_fwr}
\alias{post_hoc_est_far}
\title{Post hoc estimation for FWR and FAR}
\usage{
post_hoc_est_fwr(
  combination_method,
  A_threshold,
  B_threshold,
  C_threshold,
  baserate_hp,
  devices,
  samplesize,
  min_number = NULL,
  min_prob = NULL
)

post_hoc_est_far(
  combination_method,
  A_threshold,
  B_threshold,
  C_threshold,
  baserate_hp,
  devices,
  samplesize,
  min_number = NULL,
  min_prob = NULL
)
}
\arguments{
\item{combination_method}{Number (1 to 18) corresponding to the test method.}

\item{A_threshold}{(scalar integer) Threshold for Test A (1 to 6).}

\item{B_threshold}{(scalar integer) Threshold for Test B (1 to 6).}

\item{C_threshold}{(scalar integer) Threshold for Test C (1 to 6).}

\item{baserate_hp}{Sets the (estimated) prevalence of headphones in the
target population as a number between 0 and 1. Defaults to the unbiased
prevalence B of 0.1767 from \insertCite{HALT_2;textual}{HALT}.}

\item{devices}{Sets the desired playback device. Possible settings are
\code{"HP"} for headphones or \code{"LS"} for loudspeakers.}

\item{samplesize}{Number of participants who got a test result indicating
the use of the target device and who are therefore in the final sample.}

\item{min_number}{minimum number of participants \code{k} who are in the
final sample and should have used the correct device.
When you use \code{min_number} you cannot use \code{min_prob}, i.e.
\code{min_prob = NULL}.}

\item{min_prob}{(greater than 0, less than 1) for the event that
at least an unknown number of participants \code{k} who are in the final
sample used the correct device.
When you use \code{min_prob} you cannot use \code{min_number}, i.e.
\code{min_number = NULL}.}
}
\description{
Post hoc estimation for FWR and FAR
}
