% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prevalence_calc.R
\name{post_hoc_est_scc}
\alias{post_hoc_est_scc}
\title{Post hoc estimation for SCC}
\usage{
post_hoc_est_scc(
  combination_method,
  A_threshold,
  B_threshold,
  C_threshold,
  switch_to_target,
  devices,
  target_selfreported,
  target_tested,
  min_number = NULL,
  min_prob = NULL
)
}
\arguments{
\item{combination_method}{Number (1 to 18) corresponding to the test method.}

\item{A_threshold}{(scalar integer) Threshold for Test A (1 to 6).}

\item{B_threshold}{(scalar integer) Threshold for Test B (1 to 6).}

\item{C_threshold}{(scalar integer) Threshold for Test C (1 to 6).}

\item{switch_to_target}{Sets the (estimated) switching prevalence.
The switching prevalence describes the probability that a participant who
indicates the use of a device other than the target device actually switches
to the target device after being prompted to do so.}

\item{devices}{Sets the desired playback device. Possible settings are
\code{"HP"} for headphones or \code{"LS"} for loudspeakers.}

\item{target_selfreported}{Number of participants who reported the use of
the target device.}

\item{target_tested}{Number of participants who reported the use of a
playback device other than the target device and got a test result
indicating the use of the target device.}

\item{min_number}{minimum number of participants \code{k} who are in the
final sample and should have used the correct device.
When you use \code{min_number} you cannot use \code{min_prob}, i.e.
\code{min_prob = NULL}.}

\item{min_prob}{(greater than 0, less than 1) for the event that
at least an unknown number of participants \code{k} who are in the final
sample used the correct device.
When you use \code{min_prob} you cannot use \code{min_number}, i.e.
\code{min_number = NULL}.}
}
\description{
This function provides probabilistic statements about the composition of a
sample after application of a certain test procedure within SCC. For this
purpose a Binomial distribution is used.
}
\details{
Within SCC the final sample contains participants who reported the use of
the target playback device and participants who did not do so but got a
test result indicating the use of the target device.
Given a test procedure, switching prevalence, the number of participants who
reported the use of the target playback device, and the number of
participants who reported the use of a device other than the target device
and got a test result indicating the use of the target device the event that
at least \code{k} participants who are in the final sample used the correct
device is considered.
The function either calculates the minimum probability for this event for a
given \code{k} or \code{k} for a given probability for this event.
}
\note{
Only one of the arguments \code{min_number} and \code{min_prob} can be
used.
}
