% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_hoc_estimation.R
\name{post_hoc_explanation}
\alias{post_hoc_explanation}
\title{Generate explanation text for Post hoc estimation}
\usage{
post_hoc_explanation(
  screening_strat,
  combination_method,
  A,
  B,
  C,
  devices,
  baserate_hp,
  min_number,
  min_prob,
  sample_size = NA,
  target_selfreported = NA,
  target_tested = NA,
  switch_to_target = NA,
  min_data_qual_perc
)
}
\description{
This functions provides an explanatory text for a post hoc estimation.
It is used inside \code{\link{post_hoc_estimation}}.
}
