% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_hoc_estimation.R
\name{post_hoc_tibble}
\alias{post_hoc_tibble}
\title{Construct table for post hoc estimation}
\usage{
post_hoc_tibble(
  screening_strat,
  combination_method,
  A,
  B,
  C,
  target_device,
  sample_size = NA,
  target_selfreported = NA,
  target_tested = NA,
  baserate_hp,
  switch_to_target = NA,
  min_number,
  min_prob
)
}
\description{
This function constructs a tibble with a screening test configuration
including screening strategy and the results of a post hoc estimation.
}
