% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screening_strategies.R
\name{screening_strategies}
\alias{screening_strategies}
\title{HALT Screening Strategies}
\usage{
screening_strategies()
}
\description{
List the screening strategies available in the HALT.
For details, see \insertCite{HALT_2;textual}{HALT}.
}
\references{
\insertAllCited{}
}
