% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_config.R
\name{show_config}
\alias{show_config}
\title{Showing the properties of the automatic configuration}
\usage{
show_config(config = HALT::auto_config())
}
\arguments{
\item{config}{Output from \link{HALT}{\code{auto_config}} or an accordingly named
list.}
}
\description{
This function prints out the properties of the automatic configuration:
the True Headphones Rate, the True Loudspeakers Rate, the Headphones
Predictive Value, the Loudspeakers Predictive Value and the Utility for
maximizing percent correct.
}
