% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple-device-page.R
\name{simple_device_page}
\alias{simple_device_page}
\title{Playback device self-report page}
\usage{
simple_device_page(dict = HALT::HALT_dict)
}
\arguments{
\item{dict}{The psychTestR dictionary used for internationalisation.}
}
\description{
This function creates a page where participant report their playback device.
}
