% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prevalence_calc.R
\name{tests_scc_utility}
\alias{tests_scc_utility}
\title{Calculate overall utilities and target device probabilities for SCC}
\usage{
tests_scc_utility(
  baserate_hp = 211/1194,
  devices = "HP",
  switch_to_target = 3/4
)
}
\arguments{
\item{baserate_hp}{Sets the (estimated) prevalence of headphones in the
target population as a number between 0 and 1. Defaults to the unbiased
prevalence B of 0.1767 from \insertCite{HALT_2;textual}{HALT}.}

\item{devices}{Sets the desired playback device. Possible settings are
\code{"HP"} for headphones or \code{"LS"} for loudspeakers.}

\item{switch_to_target}{Sets the (estimated) switching prevalence.
The switching prevalence describes the probability that a participant who
indicates the use of a device other than the target device actually switches
to the target device after being prompted to do so.}
}
\description{
Calculate overall utilities and target device probabilities for SCC
}
