% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo.R
\name{HALT_demo}
\alias{HALT_demo}
\title{Demo HALT}
\usage{
HALT_demo(
  config = HALT::auto_config(screening_parts = FALSE),
  admin_password = "demo",
  researcher_email = "",
  dict = HALT::HALT_dict,
  language = "en"
)
}
\arguments{
\item{config}{(HALT_config object) A HALT_config object, generate with HATL::make_config() or auto_config() (default).}

\item{admin_password}{(Scalar character) Password for accessing the admin panel.
Defaults to \code{"demo"}.}

\item{researcher_email}{(Scalar character)
If not \code{NULL}, this researcher's email address is displayed
at the bottom of the screen so that online participants can ask for help.
Defaults to \email{longgold@gold.uc.ak},
the email address of this package's developer.}

\item{dict}{The psychTestR dictionary used for internationalisation.}

\item{language}{The language you want to run your demo in.
Possible languages include English (\code{"EN"}) and German (\code{"DE"}).
The first language is selected by default}
}
\description{
This function launches a demo for the HALT, which is just a wrapper to HALT_standalong with an no_screening flag.
This function launches a demo for the HALT, which is just a wrapper to HALT_standalong with an no_screening flag.
}
