% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_config.R
\name{make_config}
\alias{make_config}
\title{Create a configuration object for the HALT module}
\usage{
make_config(
  combination_method,
  A_threshold,
  B_threshold,
  C_threshold,
  baserate_hp,
  devices,
  use_scc = F,
  loop_exclude = 5L,
  lr_img_exclude = TRUE,
  lr_audio_exclude = TRUE,
  devices_exclude = TRUE
)
}
\arguments{
\item{combination_method}{Number (1 to 18) corresponding to the test method.}

\item{A_threshold}{(scalar integer) Threshold for Test A (1 to 6).}

\item{B_threshold}{(scalar integer) Threshold for Test B (1 to 6).}

\item{C_threshold}{(scalar integer) Threshold for Test C (1 to 6).}

\item{baserate_hp}{Sets the (estimated) prevalence of headphones in the
target population as a number between 0 and 1. Defaults to the unbiased
prevalence B of 0.1767 from \insertCite{HALTpaper;textual}{HALT}.}

\item{devices}{Sets the desired playback device. Possible settings are
\code{"HP"} for headphones or \code{"LS"} for loudspeakers.}

\item{use_scc}{(boolean, default = F) Flag whether "Split, Convince, Compare" page shall be shown.}

\item{loop_exclude}{(integer, default = 5) Number of loops for item po2.}

\item{lr_img_exclude}{(boolean, default = T) Flag, if wrong answer on left-right image question shall lead to exclusiogn.}

\item{lr_audio_exclude}{(boolean, default = T) Flag, if wrong answer on left-right audio question shall lead to exclusiogn.}

\item{devices_exclude}{(boolean, default = T) Flag, if wrong answer on left-right device question shall lead to exclusiogn.}
}
\description{
This function creates a configuration list for the psychTestR HALT function.
}
\references{
\insertRef{HALTpaper}{HALT}
}
