% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prevalence_calc.R
\name{post_hoc_est_scc}
\alias{post_hoc_est_scc}
\title{Post hoc estimation for SCC}
\usage{
post_hoc_est_scc(
  combination_method,
  A_threshold,
  B_threshold,
  C_threshold,
  switch_to_target,
  devices,
  target_selfreported,
  target_tested,
  min_number = NULL,
  min_prob = NULL
)
}
\arguments{
\item{combination_method}{Number (1 to 18) corresponding to the test method.}

\item{A_threshold}{(scalar integer) Threshold for Test A (1 to 6).}

\item{B_threshold}{(scalar integer) Threshold for Test B (1 to 6).}

\item{C_threshold}{(scalar integer) Threshold for Test C (1 to 6).}

\item{switch_to_target}{Sets the (estimated) switching prevalence.
The switching prevalence describes the probability that a participant who
indicates the use of a device other than the target device actually switches
to the target device after being prompted to do so.}

\item{devices}{Sets the desired playback device. Possible settings are
\code{"HP"} for headphones or \code{"LS"} for loudspeakers.}

\item{target_selfreported}{Number of participants who reported the use of
the target device.}

\item{target_tested}{Number of participants who reported the use of a
playback device other than the target device and got a test result
indicating the use of the target device.}
}
\description{
This function provides probabilistic statements about the composition of a
sample after application of a certain test procedure within SCC. For this
purpose a Binomial distribution is used.
}
